refactor: enhance state machine transitions and reintroduce event determination logic
